+++ nemo/collections/asr/modules/wav2vec_modules.py
@@ -22,24 +22,28 @@ from typing import Optional, Tuple
 import numpy as np
 import torch
 import torch.nn.functional as F
+from pytorch_lightning import Trainer
 from torch import nn
 
-from nemo.collections.asr.models.wav2vec.wav2vec_config import Wav2VecMaskType
+from nemo.collections.asr.models.wav2vec.wav2vec_config import Wav2VecMaskType, Wav2VecCTCEncoderConfig
+from nemo.collections.asr.models.wav2vec.wav2vec_model import Wav2VecEncoderModel
+from nemo.core import NeuralModule
+from nemo.core.neural_types import NeuralType, LossType, EncodedRepresentation, VoidType, AudioSignal, LengthsType
 
 
-class GumbelVectorQuantizer(nn.Module):
+class GumbelVectorQuantizer(NeuralModule):
     def __init__(
-        self,
-        dim,
-        num_vars,
-        temp,
-        groups,
-        combine_groups,
-        vq_dim,
-        time_first,
-        activation=nn.GELU(),
-        weight_proj_depth=1,
-        weight_proj_factor=1,
+            self,
+            dim,
+            num_vars,
+            temp,
+            groups,
+            combine_groups,
+            vq_dim,
+            time_first,
+            activation=nn.GELU(),
+            weight_proj_depth=1,
+            weight_proj_factor=1,
     ):
         """Vector quantization using gumbel softmax
 
@@ -122,6 +126,32 @@ class GumbelVectorQuantizer(nn.Module):
         z = self.vars.squeeze(0).index_select(0, indices.flatten()).view(b, n, -1)
         return z
 
+    @property
+    def input_types(self):
+        """Returns definitions of module input ports.
+        """
+        if self.time_first:
+            return {
+                "x": NeuralType(('B', 'T', 'D'), EncodedRepresentation())
+            }
+        return {
+            "x": NeuralType(('B', 'D', 'T'), EncodedRepresentation())
+        }
+
+    @property
+    def output_types(self):
+        """Returns definitions of module output ports.
+        """
+        if self.time_first:
+            return {
+                "x": NeuralType(('B', 'T', 'D'), EncodedRepresentation()),
+                "quantize_prob_ppl": NeuralType(elements_type=LossType()),
+            }
+        return {
+            "x": NeuralType(('B', 'D', 'T'), EncodedRepresentation()),
+            "quantize_prob_ppl": NeuralType(elements_type=LossType()),
+        }
+
     def forward(self, x):
 
         if not self.time_first:
@@ -165,18 +195,29 @@ class GumbelVectorQuantizer(nn.Module):
         return x, quantize_prob_ppl, cur_codebook_temp
 
 
-class TransposeLast(nn.Module):
-    def __init__(self, deconstruct_idx=None):
-        super().__init__()
-        self.deconstruct_idx = deconstruct_idx
+class TransposeLast(NeuralModule):
+    """
+    Transposes last dimension. Useful for adding to a sequential block.
+    """
 
     def forward(self, x):
-        if self.deconstruct_idx is not None:
-            x = x[self.deconstruct_idx]
         return x.transpose(-2, -1)
 
+    @property
+    def input_types(self):
+        return {
+            "x": NeuralType(elements_type=VoidType())
+        }
 
-class SamePad(nn.Module):
+    @property
+    def output_types(self):
+        return {
+            "x": NeuralType(elements_type=VoidType())
+        }
+
+
+class SamePad(NeuralModule):
+
     def __init__(self, kernel_size):
         super().__init__()
         self.remove = kernel_size % 2 == 0
@@ -186,17 +227,100 @@ class SamePad(nn.Module):
             x = x[:, :, :-1]
         return x
 
+    @property
+    def input_types(self):
+        return {
+            "x": NeuralType(('B', 'T', 'C', 'D'), EncodedRepresentation())
+        }
+
+    @property
+    def output_types(self):
+        return {
+            "x": NeuralType(('B', 'T', 'C', 'D'), EncodedRepresentation())
+        }
+
+
+class Wav2VecCTCEncoder(NeuralModule):
+    def __init__(self,
+                 wav2vec_encoder: Wav2VecEncoderModel,
+                 cfg: Wav2VecCTCEncoderConfig,
+                 encoder_dim: int,
+                 trainer: Trainer):
+        super().__init__()
+        self.trainer = trainer
+        self.final_dropout = nn.Dropout(cfg.final_dropout)
+        # Add 1 for blank char
+        self.vocabulary = cfg.vocabulary
+        self.freeze_encoder_after_steps = cfg.freeze_encoder_after_steps
+        self._num_classes = len(self.vocabulary) + 1
+        self.apply_mask = cfg.mask.apply_mask
+        self.wav2vec_encoder = wav2vec_encoder
+
+        if self.apply_mask:
+            # Override encoder mask cfg with ctc encoder mask cfg
+            self.wav2vec_encoder.mask_cfg = cfg.mask
+        self.wav2vec_encoder.remove_pretraining_modules()
+
+        self.proj = self.linear(encoder_dim, self._num_classes)
+
+    def linear(self, in_features, out_features, bias=True):
+        m = nn.Linear(in_features, out_features, bias)
+        nn.init.xavier_uniform_(m.weight)
+        if bias:
+            nn.init.constant_(m.bias, 0.0)
+        return m
+
+    def forward(self, audio_signal, audio_lengths):
+        freeze_encoder_at_step = (
+                self.freeze_encoder_after_steps is not None and self.freeze_encoder_after_steps <= self.trainer.global_step
+        )
+
+        if freeze_encoder_at_step:
+            with torch.no_grad():
+                x, padding_mask = self.wav2vec_encoder.extract_features(
+                    source=audio_signal, audio_lengths=audio_lengths, mask=self.apply_mask and self.training
+                )
+        else:
+            x, padding_mask = self.wav2vec_encoder.extract_features(
+                source=audio_signal, audio_lengths=audio_lengths, mask=self.apply_mask and self.training
+            )
+
+        x = self.final_dropout(x)
+        x = self.proj(x)
+
+        non_padding_mask = ~padding_mask
+        output_lengths = non_padding_mask.long().sum(-1)
+        return x, output_lengths
+
+    @property
+    def num_classes_with_blank(self):
+        return self._num_classes
+
+    @property
+    def input_types(self):
+        return {
+            "audio_signal": NeuralType(('B', 'T'), AudioSignal()),
+            "audio_lengths": NeuralType(tuple("B"), LengthsType())
+        }
+
+    @property
+    def output_types(self):
+        return {
+            "x": NeuralType(('B', 'T', 'D'), EncodedRepresentation()),
+            "output_lengths": NeuralType(tuple("B"), LengthsType())
+        }
+
 
 def compute_mask_indices(
-    shape: Tuple[int, int],
-    padding_mask: Optional[torch.Tensor],
-    mask_prob: float,
-    mask_length: int,
-    mask_type: Wav2VecMaskType = Wav2VecMaskType.static,
-    mask_other: float = 0.0,
-    min_masks: int = 0,
-    no_overlap: bool = False,
-    min_space: int = 0,
+        shape: Tuple[int, int],
+        padding_mask: Optional[torch.Tensor],
+        mask_prob: float,
+        mask_length: int,
+        mask_type: Wav2VecMaskType = Wav2VecMaskType.static,
+        mask_other: float = 0.0,
+        min_masks: int = 0,
+        no_overlap: bool = False,
+        min_space: int = 0,
 ) -> np.ndarray:
     """
     Computes random mask spans for a given shape
